Add new solutions to binary tree problems
This commit introduces solutions to three problems concerning binary trees. These problems and their solutions are as follow: flattening a binary tree to a linked list, calculating the sum of all root-to-leaf numbers, and implementing a level-by-level traversal of a binary tree. Each solution is contained in a separate java file.