Rename Client.java to MainClient.java
The filename for the main client class in the factory method design pattern has been refactored from "Client.java" to "MainClient.java". This was done to provide a clearer distinction and understanding of the class's role within the overall program. Additional comments were also added to explain the purpose and principles of the factory design pattern.